+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 
 v5.7.0
 ------
+- feature: +more popover for timegrid (#4218)
 - feature: eventMinHeight (#961)
 - feature: timeGridEventShortHeight
 - fix: events can be completely hidden behind others with custom eventOrder (#6019)
@@ -8,7 +9,6 @@ v5.7.0
 - fix: less homogeneous event widths in timegrid (#5004)
 
 write tests:
-- feature: +more popover for timegrid (#4218)
 - feature: +more popover for timeline (#4827)
 
 verify manually (then, write tests):